move mobile menu to separate view
